refactor(core): extract common layout handling from bootloader ui implementations in rust
[no changelog]
