+++ core/embed/rust/src/ui/layout/mod.rs
@@ -1,3 +1,9 @@
+#[cfg(feature = "micropython")]
 pub mod obj;
+
+#[cfg(feature = "micropython")]
 pub mod result;
+pub mod simplified;
+
+#[cfg(feature = "micropython")]
 pub mod util;
